Centralized compilation management and error handling.
Compilation is now managed by SourceUnit and CompilationUnit objects to which errors
are reported.  CompilerConfiguration provides a single object for controlling compiler
configuration.  CompilationFailedException is thrown for any compilation failures, and
links back to the CompilationUnit and SourceUnit that discovered the problem and that
house the error messages.  These objects all reside in org.codehaus.groovy.control.


git-svn-id: http://svn.codehaus.org/groovy/trunk/groovy/groovy-core@1082 a5544e8c-8a19-0410-ba12-f9af4593a198