GROOVY-7363: Frequent compilation error on cascading generic types (closes #158)
Reliable type resolution in the static type checker with covariant getters

Prefer a non-bridge getter method

Preferring a non-bridge getter over a bridge one ensures that the type
resolution of expressions in the static type checker always uses the
covariant return type of getters (due to generics, for example) and doesn't
lose accuracy, resulting in spurious type checking errors.

It's worth noting that the check is actually whether the method is synthetic
or not, because MethodNode doesn't keep the bridge flag, but that should be
good enough.