GROOVY-7746: Add BufferedIterator and Groovy JDK methods for it (closes #247)
BufferedIterator is a wrapper for Iterator that adds a new method
`head` to return the next item without consuming it.  A helper method
was also added to DefaultGroovyMethods to create a buffered wrapper by
calling iterator.buffered() on any Iterator.  BufferedIterator is
loosely based on the Scala class of the same name.

Updates based on pull request feedback.

* Made `BufferedIterator` interface and provide two implementations:
  one for iterators with no `remove()` support, and one for lists with
  `remove()` support.
* Added `bufferedIterator()` GDK method to `Iterable` and `List`
  returning the appropriate implementation of BufferedIterator.

Fixed documentation errors and implemented remove for IteratorBufferedIterator.