Remove duplicate ReferenceBundle definitions (closes #249)
Removed static initialization of default Soft/Weak bundles in ReferenceManager class.  The methods in that class were not used in the codebase.  Looking at the history of those static bundle references it appeared that 500 was the original threshold setting.  Typical startup creates roughly 1000 managed references, so 500 seems like a more appropriate value than 5000.